[level 1] Title: 부족한 금액 계산하기, Time: 0.07 ms, Memory: 79.4 MB -BaekjoonHub

+++ 프로그래머스/lv1/82612. 부족한 금액 계산하기/부족한 금액 계산하기.java
@@ -1,8 +1,8 @@
 class Solution {
-    public long solution(int price, int money, int count) {
+    public long solution(long price, long money, long count) {
         long answer = 0L;
         long ticket = 0L;
-        ticket = (long) price * ((long) count * ((long) count + 1))/2;
+        ticket = price * ( count * ( count + 1))/2;
         
         if (ticket - money > 0) {
             answer = ticket - money;
